requirements.txt
altgraph==0.17.2
api-display-purposes==0.0.3
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
auto-py-to-exe==2.22.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bottle==0.12.23
bottle-websocket==0.2.9
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clarifai==2.6.2
colorama==0.4.5
configparser==3.8.1
cryptography==37.0.2
EasyProcess==1.1
Eel==0.14.0
emoji==1.7.0
fake-useragent==0.1.11
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.47.0
h11==0.13.0
idna==2.10
jsonschema==2.6.0
lxml==4.9.0
MeaningCloud-python==2.0.0
numpy==1.23.0
outcome==1.2.0
pandas==1.4.3
pefile==2022.5.30
plyer==2.0.0
protobuf==3.20.1
pycodestyle==2.8.0
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
scrapper-boilerplate==0.2.5
selenium==4.2.0
semantic-version==2.10.0
setuptools-rust==1.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.38
toml==0.10.2
tornado==6.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
webdriver-manager==3.7.0
webdriverdownloader==1.1.0.3
whichcraft==0.6.1
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
